Create Queue using an Array - Menu based program
Write a program to perform the following operations using Java.
1. Read the count of elements from user to create the simple queue using array. 
2. Create the class and methods for the following and call them using a menu based program and 
call them using a loop.
a. Show the queue
b. Store the data in the queue
c. Access the data from the queue
d. Check how many elements are there in the queue.
e. Check to see if the queue is full?
f. Check to see if queue is empty